feat(router) add router_update_frequency configuration option
Prior this change the router and plugins where checked for changes
every second (starting two timers per second on each worker). The
issue #4872 reported some issues that frequent updates may cause
on proxying performance and suggested we allow configuring the value.
This PR adds `router_update_frequency` configuration option, which
still default to that same once per second. This is mostly usable with
`router_consistency=eventual`. But it will affect `strict` too (but
raising this value with `strict` will make it more likely that the
routers are rebuild on request time (making a timers bit useless,
though it may be ok with slow traffic scenarios).

Fix #4872