perf(proxy) cache plugins route and service ids (for discussion)
At the moment plugins make Kong not to work very well when database
connection is not available. Even when we do pre-warming, because we
don't do negative cache pre-warming. That in general results a DB call.

Previous attempt to fix this was to load all the plugins in memory,
but that raised concerns of consumers and plugins applied to them.
This PR changes plugin iterator to cache route and service ids. Sure
there can be a lot of routes and services too, but those are the
entities that we already cache quite extensively.

This PR is for discussion and in my testing it solves the issue reported
here: https://github.com/Kong/kong/issues/4807:

It allows me to:

1. start kong
2. stop postgres
3. proxy all those requests that were mentioned in #4807